feat: 토스트 알림 함수 적용, 산출물 페이지 오류 수정 및 팀빌딩 입력 글자 수 제한

See merge request s11-final/S11P31E203!235

+++ client/src/pages/teamBuilding/TeamBuildingDetailPage.tsx
@@ -16,6 +16,7 @@ import RecruitmentSelector from '@/features/teamBuilding/components/createTeam/r
 import { editRecruiting } from '@/features/teamBuilding/apis/editTeam/editRecruiting';
 import { Recruitment, TeamBuildingData, TeamBuildingMember, MemberDeleteStatus } from '@/features/teamBuilding/types/teamBuildingDetail/TeamBuildingDetailTypes';
 import useTeamStore from '@/features/project/stores/useTeamStore';
+import { showToast } from '@/utils/toastUtils';
 
 const initialData: TeamBuildingData = {
   postId: 0,
@@ -88,7 +89,10 @@ const TeamBuildingDetailPage = () => {
   };
 
   const handleMessageChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setMessage(e.target.value);
+    const inputValue = e.target.value;
+    if (inputValue.length <= 250) {
+      setMessage(e.target.value);
+    }
   };
 
   const filteredCandidates = isAuthor 
@@ -111,7 +115,7 @@ const TeamBuildingDetailPage = () => {
     }
 
     if (!message.trim()) {
-      alert('댓글을 입력해주세요.');
+      showToast.warn('댓글을 입력해주세요.');
       return;
     }
 
@@ -130,7 +134,7 @@ const TeamBuildingDetailPage = () => {
       })
       .catch((err) => {
         console.error(err);
-        alert('댓글 작성 중 오류가 발생했습니다. 지원은 전체 한번만 가능합니다.');
+        showToast.error('댓글 작성 중 오류가 발생했습니다. 지원은 전체 한번만 가능합니다.');
       }); 
   };
 
@@ -141,14 +145,14 @@ const TeamBuildingDetailPage = () => {
 
   const handleSaveComment = async () => {
     if (activeCommentId === null || !editedCommentContent.trim()) {
-        alert("내용을 입력해주세요.");
+        showToast.warn("내용을 입력해주세요.");
         return;
     }
 
     const editedCandidate = filteredCandidates[activeCommentId];
 
     if (!editedCandidate) {
-        alert("수정할 댓글을 찾을 수 없습니다.");
+        showToast.warn("수정할 댓글을 찾을 수 없습니다.");
         return;
     }
 
@@ -159,28 +163,33 @@ const TeamBuildingDetailPage = () => {
 
     try {
         await editComment(data.postId, editedCandidate.recruitingMemberId, params);
-        alert("댓글이 성공적으로 수정되었습니다.");
+        showToast.success("댓글이 성공적으로 수정되었습니다.");
         window.location.reload(); // 수정 후 페이지 새로고침
     } catch (error) {
-        console.error("댓글 수정 중 오류가 발생했습니다:", error);
-        alert("댓글 수정 중 문제가 발생했습니다. 다시 시도해주세요.");
+        showToast.error("댓글 수정 중 문제가 발생했습니다. 다시 시도해주세요.");
     }
   };
 
   const handleDeleteComment = (postId:string, commentId: number) => {
-    deleteComment(parseInt(postId), commentId)
-      .then((response) => {
-        console.log(response);
-        window.location.reload();
-      })
-      .catch((err) => {
-        console.error(err);
-      });
+    const userConfirmed = window.confirm("정말로 삭제하시겠습니까?");
+    
+    if (userConfirmed) {
+      deleteComment(parseInt(postId), commentId)
+        .then((response) => {
+          console.log(response);
+          window.location.reload();
+        })
+        .catch((err) => {
+          console.error(err);
+        });
+    } else {
+      console.log("User canceled the deletion.");
+    }
   };
 
   const handleNChange = (n: number) => {
     if (data.recruitedTotal > n) {
-        alert(
+        showToast.warn(
             `현재 모집된 멤버 수(${data.recruitedTotal})가 총 모집 인원(${n})을 초과할 수 없습니다.`
         );
         return;
@@ -245,14 +254,20 @@ const TeamBuildingDetailPage = () => {
   };
 
   const handleDeletePost = () => {
-    deletePost(parseInt(postId))
-      .then(() => {
-        alert('게시글이 삭제되었습니다.');
-        navigate('/team-building');
-      })
-      .catch((err) => {
-        console.error(err);
-      });
+    const userConfirmed = window.confirm("정말로 삭제하시겠습니까?");
+
+    if (userConfirmed) {
+      deletePost(parseInt(postId))
+        .then(() => {
+          showToast.success('게시글이 삭제되었습니다.');
+          navigate('/team-building');
+        })
+        .catch((err) => {
+          console.error(err);
+        });
+    } else {
+      console.log("User canceled the deletion.");
+    }
   };
 
   const handleEditToggle = () => {
@@ -262,7 +277,7 @@ const TeamBuildingDetailPage = () => {
 
       // 모집 인원 합이 총 인원을 초과하거나 부족한 경우
       if (totalMembers !== data.memberTotal) {
-          alert(
+          showToast.warn(
               `총 모집 인원(${data.memberTotal})과 세부 포지션 합(${totalMembers})이 일치하지 않습니다.`
           );
           return;
@@ -270,7 +285,7 @@ const TeamBuildingDetailPage = () => {
 
       // 모집된 멤버 수가 총 모집 인원을 초과하는 경우
       if (data.recruitedTotal > data.memberTotal) {
-          alert(
+          showToast.warn(
               `현재 모집된 멤버 수(${data.recruitedTotal})가 총 모집 인원(${data.memberTotal})을 초과할 수 없습니다.`
           );
           return;
@@ -319,7 +334,7 @@ const TeamBuildingDetailPage = () => {
       })
       .catch((err) => {
           console.error(err);
-          alert("데이터를 저장하는 중 오류가 발생했습니다. 다시 시도해주세요."); // 사용자에게 에러 메시지를 표시합니다.
+          showToast.error("데이터를 저장하는 중 오류가 발생했습니다. 다시 시도해주세요."); // 사용자에게 에러 메시지를 표시합니다.
       });
   };
 
@@ -344,7 +359,7 @@ const TeamBuildingDetailPage = () => {
       await editComment(data.postId, recruitingMemberId, params);  
       window.location.reload();
     } catch (error) {
-      alert('처리에 실패했습니다. 다시 시도해주세요.');
+      showToast.error('처리에 실패했습니다. 다시 시도해주세요.');
     }
   };
 
@@ -401,6 +416,7 @@ const TeamBuildingDetailPage = () => {
                 placeholder="지원 메시지를 입력하세요"
                 rows={4}
                 cols={50}
+                maxLength={250}
               />
               <button onClick={handleSubmitComment} className={styles.submitButton}>지원하기</button>
             </div>
@@ -422,8 +438,14 @@ const TeamBuildingDetailPage = () => {
                   {isCommentEditing && activeCommentId === index ? (
                     <textarea
                       value={editedCommentContent}
-                      onChange={(e) => setEditedCommentContent(e.target.value)}
+                      onChange={(e) => {
+                        const inputValue = e.target.value;
+                        if (inputValue.length <= 250) {
+                          setEditedCommentContent(inputValue); // 최대 250자까지만 상태 업데이트
+                        }
+                      }}
                       className={styles.editInput}
+                      maxLength={250}
                       rows={4} // 원하는 줄 수
                       cols={50} // 원하는 열 수
                     />
